MVVM structural refactor
