+++ app/src/main/java/com/bintina/goouttolunchmvvm/restaurants/list/viewmodel/ListViewModel.kt
@@ -1,4 +1,4 @@
-package com.bintina.goouttolunchmvvm.restaurants.list
+package com.bintina.goouttolunchmvvm.restaurants.list.viewmodel
 
 import androidx.lifecycle.ViewModel
 /*
