+++ app/src/main/java/com/bintina/goouttolunchmvvm/utils/Extensions.kt
@@ -1,4 +1,4 @@
-package com.bintina.goouttolunchmvvm
+package com.bintina.goouttolunchmvvm.utils
 
 fun openSearchActivity(){
